+++ index.js
@@ -1,10 +1,10 @@
 #!/usr/bin/env node
 import { program } from 'commander';
 import { download2mp3 } from './src/download2mp3.js';
-import { getDataByUrl } from './src/getDataByUrl.js';
+import { getDataByUrl } from './src/bilibili/getDataByUrl.js';
 import { createRequire } from 'module';
-import { sleep } from './src/utils.js';
-import { validateInt } from './src/validateInt.js';
+import { sleep } from './src/utils/sleep.js';
+import { validateInt } from './src/utils/validateInt.js';
 import axios from 'axios';
 
 const require = createRequire(import.meta.url);
