+++ tests/download.test.js
@@ -8,15 +8,19 @@ vi.mock('axios', () => {
   return { default: mock };
 });
 
-vi.mock('../src/getDataByUrl.js', () => ({
+vi.mock('../src/bilibili/getDataByUrl.js', () => ({
   getDataByUrl: vi.fn(),
 }));
 
-vi.mock('../src/naming.js', () => ({
+vi.mock('../src/bilibili/getPlayUrl.js', () => ({
+  getPlayUrl: vi.fn(),
+}));
+
+vi.mock('../src/bilibili/naming.js', () => ({
   getName: vi.fn(() => 'TestFile'),
 }));
 
-vi.mock('../src/progress.js', () => ({
+vi.mock('../src/download/progress.js', () => ({
   createProgressBar: vi.fn(() => ({ tick: vi.fn() })),
 }));
 
@@ -48,11 +52,12 @@ vi.mock('fs', () => ({
 }));
 
 import axios from 'axios';
-import { getDataByUrl } from '../src/getDataByUrl.js';
-import { getName } from '../src/naming.js';
-import { createProgressBar } from '../src/progress.js';
+import { getDataByUrl } from '../src/bilibili/getDataByUrl.js';
+import { getPlayUrl } from '../src/bilibili/getPlayUrl.js';
+import { getName } from '../src/bilibili/naming.js';
+import { createProgressBar } from '../src/download/progress.js';
 import * as fs from 'fs';
-import { download } from '../src/download.js';
+import { download } from '../src/download/downloadTask.js';
 
 const defaultNaming = 'TITLE-AUTHOR-DATE';
 
@@ -63,7 +68,6 @@ function mockStreamDownload(contentLength = 1024) {
       mockStream.emit('data', Buffer.alloc(512));
       mockStream.emit('data', Buffer.alloc(contentLength - 512));
       mockStream.emit('end');
-      // pipe triggers 'finish' on writeStream after source 'end'
       ws.emit('finish');
     }, 0);
     return ws;
@@ -93,9 +97,7 @@ describe('download', () => {
 
   beforeEach(() => {
     vi.clearAllMocks();
-    // Default: stat rejects (file doesn't exist)
     fs.promises.stat.mockRejectedValue(new Error('not found'));
-    // Restore createWriteStream implementation after clearAllMocks resets it
     fs.default.createWriteStream.mockImplementation(() => {
       const ws = new EventEmitter();
       ws.destroy = vi.fn();
@@ -110,11 +112,7 @@ describe('download', () => {
 
   it('should call getDataByUrl with the given url', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload();
 
     await download('https://www.bilibili.com/video/BV1test?p=1', 1, {
@@ -127,11 +125,7 @@ describe('download', () => {
 
   it('should call getName with index, derived metadata, and naming', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload();
 
     await download('https://www.bilibili.com/video/BV1test?p=1', 3, {
@@ -142,38 +136,27 @@ describe('download', () => {
     expect(getName).toHaveBeenCalledTimes(1);
     const [index, title, author, dateStr, naming] = getName.mock.calls[0];
     expect(index).toBe(3);
-    // Multi-page: title = pages[p-1].part with spaces replaced by -
     expect(title).toBe('Part-1');
     expect(author).toBe('TestAuthor');
     expect(dateStr).toMatch(/^\d{4}-\d{1,2}-\d{1,2}$/);
     expect(naming).toBe('INDEX-TITLE');
   });
 
-  it('should fetch playurl API with correct cid and bvid', async () => {
+  it('should call getPlayUrl with correct cid and bvid', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload();
 
     await download('https://www.bilibili.com/video/BV1test?p=1', 1, {
       naming: defaultNaming,
     });
 
-    expect(axios.get).toHaveBeenCalledWith(
-      'https://api.bilibili.com/x/player/playurl?fnval=80&cid=100&bvid=BV1test',
-    );
+    expect(getPlayUrl).toHaveBeenCalledWith(100, 'BV1test');
   });
 
   it('should return filename and bar', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload();
 
     const result = await download(
@@ -187,11 +170,7 @@ describe('download', () => {
 
   it('should create progress bar with correct params', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload(2048);
 
     await download('https://www.bilibili.com/video/BV1test?p=1', 2, {
@@ -201,16 +180,11 @@ describe('download', () => {
   });
 
   it('should unlink existing file before download', async () => {
-    // Simulate file already exists: stat succeeds, then unlink is called
     fs.promises.stat.mockResolvedValueOnce(undefined);
     fs.promises.unlink.mockResolvedValueOnce(undefined);
 
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload();
 
     await download('https://www.bilibili.com/video/BV1test?p=1', 1, {
@@ -221,11 +195,7 @@ describe('download', () => {
 
   it('should reject on stream error', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
 
     const mockStream = new EventEmitter();
     mockStream.pipe = vi.fn((ws) => {
@@ -249,11 +219,7 @@ describe('download', () => {
 
   it('should reject on axios request failure', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     axios.mockRejectedValueOnce(new Error('request failed'));
 
     await expect(
@@ -265,11 +231,7 @@ describe('download', () => {
 
   it('should ignore data events after stream failure', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
 
     const mockBar = { tick: vi.fn() };
     vi.mocked(createProgressBar).mockReturnValueOnce(mockBar);
@@ -278,7 +240,6 @@ describe('download', () => {
     mockStream.pipe = vi.fn((ws) => {
       setTimeout(() => {
         mockStream.emit('error', new Error('stream error'));
-        // data after error should be ignored (failed = true guard)
         mockStream.emit('data', Buffer.alloc(100));
         mockStream.emit('end');
       }, 0);
@@ -297,21 +258,14 @@ describe('download', () => {
       // expected
     }
 
-    // bar.tick should have been called for error (tick total), but NOT for post-error data/end
-    // Should not have a 'done' or 'downloading' status after error
     const statusCalls = mockBar.tick.mock.calls.filter((c) => c[1]);
-    expect(statusCalls.length).toBeLessThanOrEqual(2); // at most initial tick + error status
+    expect(statusCalls.length).toBeLessThanOrEqual(2);
   });
 
   it('should reject on writeStream error', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
 
-    // Override createWriteStream to emit 'error'
     fs.default.createWriteStream.mockImplementation(() => {
       const ws = new EventEmitter();
       ws.destroy = vi.fn();
@@ -347,11 +301,7 @@ describe('download', () => {
 
   it('should ignore finish event after stream error', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
 
     const mockBar = { tick: vi.fn() };
     vi.mocked(createProgressBar).mockReturnValueOnce(mockBar);
@@ -360,7 +310,6 @@ describe('download', () => {
     mockStream.pipe = vi.fn((ws) => {
       setTimeout(() => {
         mockStream.emit('error', new Error('stream error'));
-        // finish after error should be ignored (failed = true guard)
         ws.emit('finish');
       }, 0);
       return ws;
@@ -378,7 +327,6 @@ describe('download', () => {
       // expected
     }
 
-    // Should not have resolved successfully, so no 'done' status tick
     const doneTicks = mockBar.tick.mock.calls.filter(
       (c) => c[1] && c[1].status === 'done',
     );
@@ -395,36 +343,26 @@ describe('download', () => {
       p: 1,
     };
     vi.mocked(getDataByUrl).mockResolvedValue(singlePageData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload();
 
     await download('https://www.bilibili.com/video/BV1test?p=1', 1, {
       naming: defaultNaming,
     });
 
-    // getName is called with the video title, not page part
     const [, title] = getName.mock.calls[0];
     expect(title).toBe('Single-Video-Title');
   });
 
   it('should use page part name for multi-page videos', async () => {
     vi.mocked(getDataByUrl).mockResolvedValue(mockVideoData);
-    axios.get.mockResolvedValue({
-      data: {
-        data: { dash: { audio: [{ baseUrl: 'https://audio.url/test.m4s' }] } },
-      },
-    });
+    vi.mocked(getPlayUrl).mockResolvedValue('https://audio.url/test.m4s');
     mockStreamDownload();
 
     await download('https://www.bilibili.com/video/BV1test?p=1', 1, {
       naming: defaultNaming,
     });
 
-    // mockVideoData has p:1 and pages[0].part = 'Part 1' -> 'Part-1'
     const [, title] = getName.mock.calls[0];
     expect(title).toBe('Part-1');
   });
